break up hexgridchunk
